mathapi/requirements.txt
Django
django-rest-framework
requests